+++ Example/AppDelegate+CarPlay.swift
@@ -70,7 +70,7 @@ extension AppDelegate: CPApplicationDelegate {
 extension AppDelegate: CarPlayManagerDelegate {
     
     func carPlayManager(_ carPlayManager: CarPlayManager,
-                        navigationServiceAlong route: Route,
+                        navigationServiceAlong routeResponse: RouteResponse,
                         routeIndex: Int,
                         routeOptions: RouteOptions,
                         desiredSimulationMode: SimulationMode) -> NavigationService {
@@ -80,7 +80,7 @@ extension AppDelegate: CarPlayManagerDelegate {
             return navigationService
         }
         
-        return MapboxNavigationService(route: route,
+        return MapboxNavigationService(routeResponse: routeResponse,
                                        routeIndex: routeIndex,
                                        routeOptions: routeOptions,
                                        simulating: desiredSimulationMode)